Authentication, Google Cloud and MySQL on Cloud

+++ buy-and-sell-backend/src/routes/deleteListing.js
@@ -1,13 +1,17 @@
 import { db } from "../database";
+import * as admin from 'firebase-admin';
 
 export const deleteListingRoute = {
     method: 'DELETE',
     path: '/api/listings/{id}',
     handler: async (req, h) => {
         const {id } = req.params;
+        const token = req.headers.authtoken;
+        const user = await admin.auth().verifyIdToken(token);  
+        const userId = user.user_id;
         await db.query(
-            'DELETE from listings where id=?',
-            [id],
+            'DELETE from listings where id=? AND user_id=?',
+            [id, userId],
         );
         return { message: 'Success!' };
     }
